component renamed (index.js -> Class, index.js imports Class)
