+++ client/components-compositors/Animated/index.js
@@ -1,13 +1,1 @@
-
-import React, { Component } from 'react';
-import AnimateChildren from './children';
-
-export default Wrapped => class Animated extends Component {
-  render () {
-    return (
-      <AnimateChildren>
-        <Wrapped {...this.props} />
-      </AnimateChildren>
-    );
-  }
-};
+export default from './Animated';
