+++ client/components/ToastrContainer/index.js
@@ -1,46 +1,1 @@
-
-import React, { Component, PropTypes } from 'react';
-import Paper from 'material-ui/Paper';
-
-import styles from './styles.css';
-
-export default class ToastrContainer extends Component {
-
-  render () {
-    return (
-      <div className={styles.toastrContainer}>
-        {::this.renderToasts()}
-      </div>
-    );
-  }
-
-  renderToasts () {
-    return this.props.toastr.toasts.map(t => {
-      const removeToast = () => this.props.actions.removeToast(t.toastNo);
-
-      return (
-        <Paper
-          className={`${styles.toast} ${styles[t.type]}`}
-          zDepth={2}
-          key={t.toastNo}
-          {...this._test(`toast-${t.toastNo}`)}
-          >
-          <a className={styles.remove} onClick={removeToast}>
-            <i className='icon-trash'></i>
-          </a>
-          <span className={styles.msg}>{t.msg}</span>
-        </Paper>
-      );
-    });
-  }
-
-}
-
-ToastrContainer.propTypes = {
-  toastr: PropTypes.shape({
-    toasts: PropTypes.array.isRequired
-  }).isRequired,
-  actions: PropTypes.shape({
-    removeToast: PropTypes.func.isRequired
-  }).isRequired
-};
+export default from './ToastrContainer';
